use correct cast for object entry data

+++ testsuite/tests/nmt_hbp.c
@@ -38,7 +38,7 @@ TS_DEF_MAIN(TS_HBProd_Disable)
     uint16_t     time = 0;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
-    TS_ODAdd(CO_KEY(0x1017, 0, CO_UNSIGNED16|CO_OBJ____RW), CO_THB_PROD, (uint32_t)&time);
+    TS_ODAdd(CO_KEY(0x1017, 0, CO_UNSIGNED16|CO_OBJ____RW), CO_THB_PROD, (uintptr_t)&time);
     TS_CreateNode(&node);
                                                       /*------------------------------------------*/
     TS_Wait(&node, 1000);                             /* wait 1000ms                              */
@@ -61,7 +61,7 @@ TS_DEF_MAIN(TS_HBProd_PreOperational)
     uint16_t     time = 50;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
-    TS_ODAdd(CO_KEY(0x1017, 0, CO_UNSIGNED16|CO_OBJ____RW), CO_THB_PROD, (uint32_t)&time);
+    TS_ODAdd(CO_KEY(0x1017, 0, CO_UNSIGNED16|CO_OBJ____RW), CO_THB_PROD, (uintptr_t)&time);
     TS_CreateNode(&node);
                                                       /*------------------------------------------*/
     TS_Wait(&node, 50);                               /* wait heartbeat time                      */
@@ -86,7 +86,7 @@ TS_DEF_MAIN(TS_HBProd_Operational)
     uint16_t     time = 50;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
-    TS_ODAdd(CO_KEY(0x1017, 0, CO_UNSIGNED16|CO_OBJ____RW), CO_THB_PROD, (uint32_t)&time);
+    TS_ODAdd(CO_KEY(0x1017, 0, CO_UNSIGNED16|CO_OBJ____RW), CO_THB_PROD, (uintptr_t)&time);
     TS_CreateNode(&node);
                                                       /*------------------------------------------*/
     TS_NMT_SEND(0x01, 1);                            /* set node-id 0x01 to operational          */
